refactor: split internal/ into core/ and ui/
internal/ blocked external imports (a Go-stdlib convention) and lumped
domain logic in with the TUI. The rename makes the dependency direction
explicit at a glance: cmd → ui → core, and core/ stays free of any UI
imports so it can be tested or reused under a different frontend.

i18n stays under core/ because core/data and core/game both depend on
it (LocalName accessors, log message keys, save-persisted Lang).
