+++ core/data/gpus.json
