+++ core/data/rooms.json
