+++ core/game/greed.go
