+++ core/game/legacy.go
