+++ core/game/research_test.go
