+++ core/game/state_test.go
